Rename SIMPLE_FUNCTION_FORMAT_SPEC → NAME_PARENTHESISED_ARGS
